Geo_layout_codebase/Geo_LayoutLM-geo_code_latest_timeoptm_version_updated_apr22_sharing_backend/inference_main/code/requirements.txt
absl-py==2.1.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
async-timeout==4.0.2
attrs==23.1.0
black==21.9b0
boto3==1.28.20
botocore==1.31.20
cachetools==5.5.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
cmake==3.27.1
exceptiongroup==1.2.2
fastapi==0.115.3
filelock==3.12.2
frozenlist==1.4.0
fsspec==2023.6.0
future==0.18.3
fuzzywuzzy==0.18.0
google-api-core==2.21.0
google-auth==2.35.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.65.0
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
huggingface-hub==0.16.4
idna==3.4
imagesize==1.4.1
isort==5.9.3
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
lightning-utilities==0.9.0
lit==16.0.6
Markdown==3.7
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.9.1
nptyping==1.4.2
numpy==1.26.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.10.0.84
overrides==3.1.0
packaging==24.1
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.10.0
proto-plus==1.25.0
protobuf==5.28.2
psutil==6.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyDeprecate==0.3.1
python-dateutil==2.9.0.post0
pytorch-lightning==1.5.10
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
rsa==4.9
s3transfer==0.6.1
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.10.1
seqeval==1.2.2
six==1.16.0
sniffio==1.3.1
starlette==0.41.1
sympy==1.12
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
timm==0.9.5
tokenizers==0.13.3
tomli==1.2.3
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchmetrics==1.0.2
torchvision==0.15.2+cu118
tqdm==4.65.0
transformers==4.31.0
triton==2.0.0
typing_extensions==4.9.0
typish==1.9.3
urllib3==1.26.16
uvicorn==0.32.0
Werkzeug==3.0.4
yarl==1.9.2
zipp==3.16.2
